+++ config/medium/Mosaic/code/CGCode.js
@@ -1,0 +1,11 @@
+const n = parseInt(readline());
+for (let i = 0; i < n; i++) {
+    const line = readline();
+}
+for (let i = 0; i < n; i++) {
+
+    // Write an answer using console.log()
+    // To debug: console.error('Debug messages...');
+
+    console.log('solved line');
+}