+++ config/medium/Mosaic/code/CGCode.php
@@ -1,0 +1,14 @@
+<?php
+fscanf(STDIN, "%d", $n);
+for ($i = 0; $i < $n; $i++)
+{
+    fscanf(STDIN, "%s", $line);
+}
+for ($i = 0; $i < $n; $i++)
+{
+
+    // Write an answer using echo(). DON'T FORGET THE TRAILING \n
+    // To debug: error_log(var_export($var, true)); (equivalent to var_dump)
+
+    echo("solved line\n");
+}