+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "ASCII Art with Logo Language".
 - configuration for "ASCII Art with Logo Language, Part 2".
 - configuration for "Rummikub 1".
+- configuration for "Mosaic".
 
 ## [1.6.0] - 2024-09-02
 ### Added